pyproject.toml
[project]
name = "emotion-detector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "google-generativeai>=0.8.5",
    "jupyterlab>=4.4.5",
    "pillow>=11.3.0",
]
